Implementado Banco de Dados MySQL
Foi criado uma classe com as funções
- Atualização da lista na memória com base no banco de dados;
- Adição de item no Banco de dados;
- Remoção de item com base no nome;

AINDA FALTA: Retirar quantidades do produto no banco de dados no caso do produto não acabar

+++ PDV/formulario/src/main/java/com/propostapdv/formulario/Produto.java
@@ -2,10 +2,10 @@ package com.propostapdv.formulario;
 
 public class Produto {
     public String Nome;
-    public int Preco;
+    public double Preco;
     public int Quantidade;
 
-    public Produto(String nome, int preco, int quantidade){
+    public Produto(String nome, double preco, int quantidade){
         Nome = nome;
         Preco = preco;
         Quantidade = quantidade;
